+++ src/components/dynamic/admin/dashboards/Participants.jsx
@@ -1,5 +1,5 @@
 "use client";
-import Dashboard from "../Dashboard.jsx";
+import Dashboard from "./Dashboard.jsx";
 import {
   FILTERS,
   HEADERS,
