+++ src/components/dynamic/admin/dashboards/Teams.jsx
@@ -1,5 +1,5 @@
 "use client";
-import Dashboard from "../Dashboard.jsx";
+import Dashboard from "./Dashboard.jsx";
 import { FILTERS, HEADERS, TAGS, STATUSES } from "@/data/dynamic/admin/Teams";
 
 const Teams = () => {
