+++ src/components/dynamic/admin/dashboards/Volunteers.jsx
@@ -1,5 +1,5 @@
 "use client";
-import Dashboard from "../Dashboard.jsx";
+import Dashboard from "./Dashboard.jsx";
 import {
   FILTERS,
   HEADERS,
