+++ src/data/dynamic/admin/Admins.js
@@ -62,9 +62,6 @@ export const columns = [
       <Checkbox
         toggle={table.getIsAllRowsSelected()}
         onClick={table.getToggleAllRowsSelectedHandler()}
-        // {...{
-        //   indeterminate: table.getIsSomeRowsSelected(),
-        // }}
       />
     ),
     cell: ({ row }) => (
